# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Self-Hosted Paymaster API
  version: 0.1.0
  description: |
    Sponsorship, admin, and health API for the self-hosted ERC-4337 (EntryPoint v0.7) paymaster.

    All quantities cross the wire as **hex or decimal strings**, never JSON numbers: fee and gas
    fields routinely exceed 2^53 and a JSON number would lose precision.

    Authentication is a bearer API key (`Authorization: Bearer pm_live_...`, or `X-API-Key`). Every
    authentication failure returns an identical `401` with no indication of which check failed.
  license:
    name: MIT
    identifier: MIT

servers:
  - url: http://localhost:3000
    description: Local development

security:
  - ApiKey: []

tags:
  - name: sponsorship
    description: Public sponsorship endpoint.
  - name: admin
    description: Policy, key, and reporting management. Requires elevated permissions.
  - name: health
    description: Liveness and readiness probes.

paths:
  /paymaster/sponsor:
    post:
      tags: [sponsorship]
      summary: Sponsor a UserOperation
      description: |
        Evaluates the configured policy and, if allowed, returns a signed `paymasterAndData` binding
        the paymaster to this exact operation for a validity window. The operation must already carry
        its gas limits — the sponsorship is priced from them, and a sponsorship over different limits
        than are finally submitted is rejected on-chain.

        Requires the `sponsor:create` permission.
      operationId: sponsor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SponsorRequest" }
      responses:
        "201":
          description: Sponsorship issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SponsorResponse" }
        "400":
          description: Malformed request, unknown chain, or unknown policy
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Authenticated but not permitted, or policy denied sponsorship
          content: { application/json: { schema: { $ref: "#/components/schemas/DenialError" } } }
        "429":
          description: A quota or spend cap is exhausted (retryable)
          content: { application/json: { schema: { $ref: "#/components/schemas/DenialError" } } }
        "503":
          description: The requested chain is configured but disabled

  /health/live:
    get:
      tags: [health]
      summary: Liveness
      description: The process is running. Checks nothing external; safe for a restart probe.
      security: []
      operationId: live
      responses:
        "200":
          description: Alive
          content:
            application/json:
              schema:
                type: object
                properties: { status: { const: ok } }

  /health/ready:
    get:
      tags: [health]
      summary: Readiness
      description: |
        Whether the service can serve a sponsorship right now (policies loaded, at least one chain
        reachable). Returns `503` when not ready, without killing the process — for a load-balancer
        probe, not a restart probe.
      security: []
      operationId: ready
      responses:
        "200": { description: Ready, content: { application/json: { schema: { $ref: "#/components/schemas/Readiness" } } } }
        "503": { description: Not ready, content: { application/json: { schema: { $ref: "#/components/schemas/Readiness" } } } }

  /admin/policies:
    get:
      tags: [admin]
      summary: List policies
      description: Requires `policy:read`.
      operationId: listPolicies
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies: { type: array, items: { $ref: "#/components/schemas/Policy" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [admin]
      summary: Create or replace a policy, then hot-reload
      description: |
        Upserts a policy and reloads immediately, so the change is live without a restart. A rule
        config that cannot build a working rule is rejected (`400`) and nothing is written.
        Requires `policy:write`.
      operationId: upsertPolicy
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Policy" } } }
      responses:
        "201": { description: Upserted, content: { application/json: { schema: { $ref: "#/components/schemas/Policy" } } } }
        "400": { description: Invalid rule config, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /admin/policies/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [admin]
      summary: Get a policy
      description: Requires `policy:read`.
      operationId: getPolicy
      responses:
        "200": { description: Policy, content: { application/json: { schema: { $ref: "#/components/schemas/Policy" } } } }
        "404": { description: Not found }
    delete:
      tags: [admin]
      summary: Delete a policy
      description: |
        Fails with `409` if an API key is pinned to the policy — deleting it would unpin those keys.
        Requires `policy:write`.
      operationId: deletePolicy
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }
        "409": { description: Policy still referenced by an API key }

  /admin/policies/reload:
    post:
      tags: [admin]
      summary: Reload the policy set on demand
      description: Requires `policy:write`.
      operationId: reloadPolicies
      responses:
        "200":
          description: Reloaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  generation: { type: integer }

  /admin/keys:
    get:
      tags: [admin]
      summary: List API keys (never returns secrets)
      description: Requires `key:read`.
      operationId: listKeys
      responses:
        "200":
          description: Keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys: { type: array, items: { $ref: "#/components/schemas/ApiKeyView" } }
    post:
      tags: [admin]
      summary: Mint an API key (returns the secret exactly once)
      description: The only response that ever contains a key secret. Requires `key:write`.
      operationId: createKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, roles]
              properties:
                name: { type: string }
                roles: { type: array, items: { enum: [sponsor, viewer, admin] } }
                policyId: { type: string }
                environment: { enum: [live, test], default: live }
                expiresAt: { type: integer, description: Unix seconds }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiKeyView" }
                  - type: object
                    properties:
                      secret: { type: string, description: Shown once; not recoverable }
                      warning: { type: string }

  /admin/keys/{id}:
    delete:
      tags: [admin]
      summary: Revoke an API key
      description: Revocation is a flag, not a delete — history survives. Requires `key:write`.
      operationId: revokeKey
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked }
        "404": { description: No active key with that id }

  /admin/sponsorships:
    get:
      tags: [admin]
      summary: List issued attestations
      description: |
        These are **commitments, not spend** — most never execute on-chain, so summing them
        overstates cost. Requires `metrics:read`.
      operationId: listSponsorships
      parameters:
        - { name: apiKeyId, in: query, schema: { type: string } }
        - { name: chainId, in: query, schema: { type: integer } }
        - { name: sender, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 1000 } }
      responses:
        "200": { description: Attestations }

  /admin/audit:
    get:
      tags: [admin]
      summary: List audit-log entries
      description: Append-only; credential values are redacted at write time. Requires `metrics:read`.
      operationId: listAudit
      parameters:
        - { name: actor, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: integer, description: Unix seconds } }
        - { name: limit, in: query, schema: { type: integer, maximum: 1000 } }
      responses:
        "200": { description: Entries }

  /admin/funding:
    get:
      tags: [admin]
      summary: Where to fund this tenant, and what it holds
      description: |
        Per multi-tenant chain: the paymaster address, this tenant's `tenantKey` (the argument to
        `depositFor`), the on-chain balance, and who may withdraw it. Always the CALLER'S own tenant,
        even for a platform operator — this names where money can be sent. Chains running the
        single-tenant paymaster are omitted. Requires `key:read`.
      operationId: listFunding
      responses:
        "200":
          description: Funding per chain
          content:
            application/json:
              schema:
                type: object
                properties:
                  funding:
                    type: array
                    items:
                      type: object
                      properties:
                        chainId: { type: integer }
                        chainName: { type: string }
                        paymaster: { type: string }
                        tenantKey: { type: string, description: keccak256 of the tenant id }
                        balanceWei: { type: string, nullable: true, description: null when the chain did not answer }
                        controller:
                          type: object
                          nullable: true
                          properties:
                            address: { type: string, description: the zero address when nobody controls it yet }
                            activeAt: { type: integer, description: Unix seconds from which it may withdraw }
                            active: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /admin/funding/controller:
    post:
      tags: [admin]
      summary: Attest that a wallet may withdraw this tenant's balance
      description: |
        Returns a signed `ControllerAssignment` for the named wallet, which the CUSTOMER then submits
        to `TenantPaymaster.claimController`. No operator transaction is involved.

        Refused with `409` when the balance already has a controller: replacing one is that
        controller's own call (`transferController`) or the owner's, never the online signer's. The
        claimed wallet cannot withdraw for `CONTROLLER_CLAIM_DELAY` (one day), during which the owner
        can void the claim. Requires `funding:write`.
      operationId: issueControllerAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chainId, controller]
              properties:
                chainId: { type: integer }
                controller: { type: string, description: the wallet that will withdraw }
      responses:
        "200":
          description: A signed assignment, valid for 15 minutes
          content:
            application/json:
              schema:
                type: object
                properties:
                  chainId: { type: integer }
                  paymaster: { type: string }
                  tenantKey: { type: string }
                  controller: { type: string }
                  nonce: { type: string }
                  deadline: { type: integer, description: Unix seconds }
                  signature: { type: string }
        "400": { description: The chain holds no per-tenant balances }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { description: The balance already has a controller }

  /admin/billing:
    get:
      tags: [admin]
      summary: Plans, the treasury address, and this account's payment reference
      description: |
        `paymentReference` is what a payment transfer must carry as its transaction data for this
        account to be able to claim it. Requires `key:read`, and readable at every subscription state
        — a customer who cannot see what to pay cannot pay it.
      operationId: getBillingOptions
      responses:
        "200":
          description: What is for sale
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        periodSeconds: { type: integer }
                        priceWei: { type: object, additionalProperties: { type: string }, description: wei per period, keyed by chain id }
                        limits:
                          type: object
                          properties:
                            operationsPerDay: { type: string, nullable: true }
                            weiPerDay: { type: string, nullable: true }
                        chainIds: { type: array, nullable: true, items: { type: integer } }
                  treasury: { type: string, nullable: true }
                  paymentReference: { type: string, nullable: true }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /admin/subscription:
    get:
      tags: [admin]
      summary: This tenant's subscription and payment history
      description: |
        State is DERIVED from `paidThrough` and the clock, never stored. `renewalDue` marks the notice
        window. Requires `key:read`.
      operationId: getSubscription
      responses:
        "200":
          description: Subscription state
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: object
                    properties:
                      state: { type: string, enum: [active, grace, lapsed, none] }
                      plan: { type: string }
                      paidThrough: { type: integer }
                      graceEndsAt: { type: integer }
                      allowsSponsorship: { type: boolean }
                      renewalDue: { type: boolean }
                  payments: { type: array, items: { type: object } }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /admin/subscription/claim:
    post:
      tags: [admin]
      summary: Credit a subscription payment made on chain
      description: |
        The customer sends the plan price to the treasury with their `paymentReference` as the
        transaction data, then claims it here by hash. Amount, recipient, account and confirmations are
        all read from the chain; nothing in this request decides what was bought. One transaction buys
        one period however many times it is claimed. Requires `funding:write`.
      operationId: claimSubscriptionPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chainId, txHash, planId]
              properties:
                chainId: { type: integer }
                txHash: { type: string }
                planId: { type: string }
      responses:
        "201": { description: Credited }
        "400": { description: Unknown plan }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { description: That payment was already credited }
        "422":
          description: |
            The transaction does not pay for this: `NOT_FOUND`, `REVERTED`, `WRONG_RECIPIENT`,
            `WRONG_ACCOUNT`, `UNDERPAID`, `UNCONFIRMED` or `NOT_SOLD_ON_CHAIN` (in `code`).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }

  /admin/subscriptions/payments:
    post:
      tags: [admin]
      summary: Record a payment against any tenant (platform only)
      description: |
        The one write that crosses the tenant boundary, because billing is something the platform does
        TO an account. Requires `billing:write`, held only by the `platform` role. Customers use
        `/admin/subscription/claim` instead.
      operationId: recordSubscriptionPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, plan, periodSeconds]
              properties:
                tenantId: { type: string }
                plan: { type: string }
                periodSeconds: { type: integer, description: added to max(now, paidThrough) }
                amountWei: { type: string }
                chainId: { type: integer }
                txHash: { type: string }
                note: { type: string }
      responses:
        "201": { description: Recorded }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { description: That transaction was already recorded }

components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      description: "Authorization: Bearer pm_live_...  (or the X-API-Key header)"

  responses:
    Unauthorized:
      description: Invalid or missing API key (uniform for all auth failures)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }

  schemas:
    Quantity:
      type: string
      description: A uint256 as a hex (0x...) or decimal string. Never a JSON number.
      examples: ["0x4a817c800", "20000000000"]
    Address:
      type: string
      pattern: "^0x[0-9a-fA-F]{40}$"
    Hex:
      type: string
      pattern: "^0x[0-9a-fA-F]*$"

    SponsorRequest:
      type: object
      required: [chainId, userOperation]
      properties:
        chainId: { type: integer }
        policyId:
          type: string
          description: Ignored when the API key pins its own policy.
        userOperation:
          type: object
          required: [sender, nonce, callData, callGasLimit, verificationGasLimit, preVerificationGas, maxFeePerGas, maxPriorityFeePerGas]
          properties:
            sender: { $ref: "#/components/schemas/Address" }
            nonce: { $ref: "#/components/schemas/Quantity" }
            factory: { $ref: "#/components/schemas/Address" }
            factoryData: { $ref: "#/components/schemas/Hex" }
            callData: { $ref: "#/components/schemas/Hex" }
            callGasLimit: { $ref: "#/components/schemas/Quantity" }
            verificationGasLimit: { $ref: "#/components/schemas/Quantity" }
            preVerificationGas: { $ref: "#/components/schemas/Quantity" }
            maxFeePerGas: { $ref: "#/components/schemas/Quantity" }
            maxPriorityFeePerGas: { $ref: "#/components/schemas/Quantity" }
            signature: { $ref: "#/components/schemas/Hex" }

    SponsorResponse:
      type: object
      description: |
        The v0.7 unpacked paymaster fields to put on the operation. `paymasterAndData` is the same
        bytes concatenated, for v0.6-shaped clients. The paymaster's signature is the tail of
        paymasterData — it is deliberately not returned as a separate field.
      properties:
        paymaster: { $ref: "#/components/schemas/Address" }
        paymasterVerificationGasLimit: { $ref: "#/components/schemas/Hex" }
        paymasterPostOpGasLimit: { $ref: "#/components/schemas/Hex" }
        paymasterData: { $ref: "#/components/schemas/Hex" }
        paymasterAndData: { $ref: "#/components/schemas/Hex" }
        validUntil: { type: integer, description: Unix seconds }
        validAfter: { type: integer, description: Unix seconds }
        expiresAt: { type: string, format: date-time }
        metadata:
          type: object
          properties:
            chainId: { type: integer }
            policyId: { type: string }
            signer: { $ref: "#/components/schemas/Address" }
            maxCost: { type: string, description: "Worst-case cost committed, in wei" }
            entryPoint: { $ref: "#/components/schemas/Address" }

    Policy:
      type: object
      required: [id, name, rules]
      properties:
        id: { type: string, pattern: "^[a-zA-Z0-9._:-]{1,128}$" }
        name: { type: string }
        description: { type: string }
        enabled: { type: boolean, default: true }
        rules:
          type: array
          items:
            type: object
            required: [ruleType]
            properties:
              ruleType:
                enum: [chain-enabled, sender-allowlist, sender-blocklist, target-allowlist, method-allowlist, no-value-transfer, quota]
              config:
                type: object
                description: Shape depends on ruleType; validated server-side per type.

    ApiKeyView:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        displayPrefix: { type: string }
        roles: { type: array, items: { type: string } }
        policyId: { type: string }
        enabled: { type: boolean }
        createdAt: { type: integer }
        expiresAt: { type: integer }
        lastUsedAt: { type: integer }

    Readiness:
      type: object
      properties:
        status: { enum: [ready, not_ready] }
        policies:
          type: object
          properties:
            loaded: { type: boolean }
            generation: { type: integer }
        chains:
          type: array
          items:
            type: object
            properties:
              chainId: { type: integer }
              healthy: { type: boolean }
              blockNumber: { type: string }
              latencyMs: { type: integer }
              error: { type: string }

    Error:
      type: object
      properties:
        error: { type: string, description: Stable machine-readable code }
        message: { type: string }

    DenialError:
      type: object
      description: A policy denial. Carries the stable code, never the internal reason.
      properties:
        error: { const: SPONSORSHIP_DENIED }
        code:
          enum: [CHAIN_DISABLED, SENDER_NOT_ALLOWED, SENDER_BLOCKED, TARGET_NOT_ALLOWED, METHOD_NOT_ALLOWED, CALLDATA_UNDECODABLE, QUOTA_EXCEEDED, SPEND_CAP_EXCEEDED, VALUE_NOT_ALLOWED, RULE_ERROR]